requirements.txt
Flask==1.1.2
pymongo==3.12.0
python_bcrypt==0.3.2
gunicorn==20.1.0
